allow methlyation download to resume

+++ transform/tcga/download_methylation.py
@@ -1,5 +1,6 @@
 import argparse
 import json
+import os
 import requests
 import sys
 
@@ -25,80 +26,100 @@ if __name__ == "__main__":
         action="store_true",
         help="write file manifest, but do not download the files"
     )
+    parser.add_argument(
+        "--force",
+        default=False,
+        action="store_true",
+        help="force download the files, don't resume a partial download"
+    )
     args = parser.parse_args()
-    output_name = "source/tcga/methylation/" + args.platform.replace(" ", "")
+    output_name = os.path.join("source/tcga/methylation/", args.platform.replace(" ", ""))
+    if not os.path.isdir("source/tcga/methylation"):
+        os.mkdirs("source/tcga/methylation")
 
-    query = {
-        "op": "and",
-        "content": [
-            {
-                "op": "=",
-                "content": {
-                    "field": "data_type",
-                    "value": [
-                        "Methylation beta value"
-                    ]
+    output_archive = output_name + ".tar.gz"
+    if os.path.exists(output_archive) and not args.force:
+        print('found archive...')
+        sys.exit(0)
+
+    output_manifest = output_name + ".map"
+    if not os.path.exists(output_manifest) or args.force:
+        query = {
+            "op": "and",
+            "content": [
+                {
+                    "op": "=",
+                    "content": {
+                        "field": "data_type",
+                        "value": [
+                            "Methylation beta value"
+                        ]
+                    }
+                },
+                {
+                    "op": "=",
+                    "content": {
+                        "field": "cases.project.program.name",
+                        "value": [
+                            "TCGA"
+                        ]
+                    }
+                },
+                {
+                    "op": "=",
+                    "content": {
+                        "field": "platform",
+                        "value": [
+                            args.platform
+                        ]
+                    }
                 }
-            },
-            {
-                "op": "=",
-                "content": {
-                    "field": "cases.project.program.name",
-                    "value": [
-                        "TCGA"
-                    ]
-                }
-            },
-            {
-                "op": "=",
-                "content": {
-                    "field": "platform",
-                    "value": [
-                        args.platform
-                    ]
-                }
-            }
-        ]
-    }
+            ]
+        }
 
-    data = {}
-    id_map = {}
-    params = {}
-    params['filters'] = json.dumps(query)
-    params['expand'] = "cases.samples,cases.project"
+        data = {}
+        id_map = {}
+        params = {}
+        params['filters'] = json.dumps(query)
+        params['expand'] = "cases.samples,cases.project"
 
-    print('querying GDC API...')
+        print('querying GDC API...')
 
-    while 'size' not in params or \
-          data['pagination']['page'] < data['pagination']['pages']:
+        while 'size' not in params or \
+              data['pagination']['page'] < data['pagination']['pages']:
 
-        params['size'] = 1000
-        req = requests.get(URL_BASE + "files", params=params)
-        data = req.json()['data']
-        for i in data['hits']:
-            for case in i["cases"]:
-                if "samples" in case:
-                    for j in case["samples"]:
+            params['size'] = 1000
+            req = requests.get(URL_BASE + "files", params=params)
+            data = req.json()['data']
+            for i in data['hits']:
+                for case in i["cases"]:
+                    if "samples" in case:
+                        for j in case["samples"]:
+                            id_map[i['id']] = {
+                                "sample": j['sample_id'],
+                                "project": case["project"]["project_id"]
+                            }
+                    else:
                         id_map[i['id']] = {
-                            "sample": j['sample_id'],
                             "project": case["project"]["project_id"]
                         }
-                else:
-                    id_map[i['id']] = {
-                        "project": case["project"]["project_id"]
-                    }
-        params['from'] = data['pagination']['from'] + \
-            data['pagination']['count']
+            params['from'] = data['pagination']['from'] + \
+                data['pagination']['count']
 
-        print('processed page', data['pagination']['page'])
+            print('processed page', data['pagination']['page'])
 
-    print('creating file manifest...')
-    output_manifest = output_name + ".map"
-    with open(output_manifest, "w") as handle:
-        handle.write("file_id\tproject_id\tsample_id\n")
-        for k, v in id_map.items():
-            handle.write("%s\t%s\t%s\n" %
-                         (k, v["project"], v.get("sample", "")))
+        print('creating file manifest...')
+        with open(output_manifest, "w") as handle:
+            handle.write("file_id\tproject_id\tsample_id\n")
+            for k, v in id_map.items():
+                handle.write("%s\t%s\t%s\n" %
+                             (k, v["project"], v.get("sample", "")))
+    else:
+        print('found existing file manifest...')
+        id_map = {}
+        with open(output_manifest, "r") as fh:
+            for line in fh:
+                id_map[line.split("\t")[0]] = None
 
     if args.manifest_only:
         sys.exit(0)
@@ -113,18 +134,21 @@ if __name__ == "__main__":
     print('downloading files...')
     headers = {'Content-type': 'application/json'}
     for index, ids in enumerate(keychunks):
-        r = requests.post(URL_BASE + 'data',
-                          data=json.dumps({"ids": ids}),
-                          headers=headers,
-                          stream=True)
         path = output_name + "-" + str(index)
         paths.append(path)
-        with open(path, 'wb') as f:
-            for chunk in r.iter_content(1024):
-                f.write(chunk)
+        if not os.path.exists(path) or args.force:
+            print('downloading chunk', str(index) + '...')
+            r = requests.post(URL_BASE + 'data',
+                              data=json.dumps({"ids": ids}),
+                              headers=headers,
+                              stream=True)
+            with open(path, 'wb') as f:
+                for chunk in r.iter_content(1024):
+                    f.write(chunk)
+        else:
+            print('found chunk', str(index) + '...')
 
     print('creating archive...')
-
     random = str(uuid4())
     archive = "archive" + "-" + random
     manifest = "manifest" + "-" + random
@@ -135,7 +159,6 @@ if __name__ == "__main__":
         call(["rm", "-f", path])
         call(["mv", archive + "/MANIFEST.txt", manifest + "/" + str(index)])
 
-    output_archive = output_name + ".tar.gz"
     call("cat" + manifest + "/* > " + archive + "/MANIFEST.txt", shell=True)
     tar = "cd " + archive + " && " + "tar czvf ../" + output_archive + \
         " . && cd .."
